move URLBuilder.
It turns out to be more useful than anticipated, so moving it out of the tests dir.
