+++ dss/util/__init__.py
@@ -1,4 +1,4 @@
-from urllib.parse import urlparse, urlencode, parse_qs
+from urllib.parse import SplitResult, parse_qs, urlencode, urlparse, urlunsplit
 
 import boto3
 from botocore.auth import SigV4Auth
@@ -59,3 +59,33 @@ def connect_elasticsearch(elasticsearch_endpoint, logger) -> Elasticsearch:
     except Exception as ex:
         logger.error("Unable to connect to Elasticsearch endpoint %s. Exception: %s", elasticsearch_endpoint, ex)
         raise ex
+
+
+class UrlBuilder:
+    def __init__(self):
+        self.splitted = SplitResult("", "", "", "", "")
+        self.query = list()
+
+    def set(self, scheme: str=None, netloc: str=None, path: str=None, fragment: str=None) -> "UrlBuilder":
+        kwargs = dict()
+        if scheme is not None:
+            kwargs['scheme'] = scheme
+        if netloc is not None:
+            kwargs['netloc'] = netloc
+        if path is not None:
+            kwargs['path'] = path
+        if fragment is not None:
+            kwargs['fragment'] = fragment
+        self.splitted = self.splitted._replace(**kwargs)
+
+        return self
+
+    def add_query(self, query_name: str, query_value: str) -> "UrlBuilder":
+        self.query.append((query_name, query_value))
+
+        return self
+
+    def __str__(self) -> str:
+        result = self.splitted._replace(query=urlencode(self.query, doseq=True))
+
+        return urlunsplit(result)
