+++ tests/infra.py
@@ -1,11 +1,10 @@
+import functools
 import json
 import logging
 import os
 import re
 import typing
-import urllib.parse
 
-import functools
 from flask import wrappers
 
 
@@ -80,33 +79,3 @@ class DSSAsserts:
             return super(DSSAsserts, self).__getattr__(item)  # type: ignore
         else:
             raise AttributeError(item)
-
-
-class UrlBuilder:
-    def __init__(self):
-        self.splitted = urllib.parse.SplitResult("", "", "", "", "")
-        self.query = list()
-
-    def set(self, scheme: str=None, netloc: str=None, path: str=None, fragment: str=None) -> "UrlBuilder":
-        kwargs = dict()
-        if scheme is not None:
-            kwargs['scheme'] = scheme
-        if netloc is not None:
-            kwargs['netloc'] = netloc
-        if path is not None:
-            kwargs['path'] = path
-        if fragment is not None:
-            kwargs['fragment'] = fragment
-        self.splitted = self.splitted._replace(**kwargs)
-
-        return self
-
-    def add_query(self, query_name: str, query_value: str) -> "UrlBuilder":
-        self.query.append((query_name, query_value))
-
-        return self
-
-    def __str__(self) -> str:
-        result = self.splitted._replace(query=urllib.parse.urlencode(self.query, doseq=True))
-
-        return urllib.parse.urlunsplit(result)
