+++ tests/test_file.py
@@ -17,7 +17,8 @@ sys.path.insert(0, pkg_root) # noqa
 
 import dss
 from dss.config import BucketStage, override_bucket_config
-from tests.infra import DSSAsserts, UrlBuilder, get_env
+from dss.util import UrlBuilder
+from tests.infra import DSSAsserts, get_env
 
 
 class TestFileApi(unittest.TestCase, DSSAsserts):
